Move mixings file
